int static casts -> compiler requested
[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1562405]

+++ Eve/SpaceObject/Children/Behaviors/Allign.cpp
@@ -71,7 +71,7 @@ std::vector<Vector3> Allign::CalculateBehavior(std::vector<DroneAgent>& agents, 
 
 		for ( auto agent = agents.begin(); agent != agents.end(); ++agent, c++ )
 		{
-			if ( c >= m_lastPullForces.size() )
+			if ( c >= static_cast< int >( m_lastPullForces.size() ) )
 			{
 				break;
 			}
